new useAccountPurgatory() hook
