+++ src/components/App.tsx
@@ -4,9 +4,10 @@ import Header from './organisms/Header'
 import Styles from '../global/Styles'
 import styles from './App.module.css'
 import { useSiteMetadata } from '../hooks/useSiteMetadata'
-import { useOcean } from '../providers/Ocean'
 import Alert from './atoms/Alert'
 import { graphql, PageProps, useStaticQuery } from 'gatsby'
+import { useAccountPurgatory } from '../hooks/useAccountPurgatory'
+import { useWeb3 } from '../providers/Web3'
 
 const contentQuery = graphql`
   query AppQuery {
@@ -35,10 +36,8 @@ export default function App({
   const purgatory = data.purgatory.edges[0].node.childContentJson.account
 
   const { warning } = useSiteMetadata()
-  const {
-    isInPurgatory: isAccountInPurgatory,
-    purgatoryData: accountPurgatory
-  } = useOcean()
+  const { accountId } = useWeb3()
+  const { isInPurgatory, purgatoryData } = useAccountPurgatory(accountId)
 
   return (
     <Styles>
@@ -47,10 +46,10 @@ export default function App({
         {(props as PageProps).uri === '/' && (
           <Alert text={warning} state="info" />
         )}
-        {isAccountInPurgatory && (
+        {isInPurgatory && (
           <Alert
             title={purgatory.title}
-            badge={`Reason: ${accountPurgatory?.reason}`}
+            badge={`Reason: ${purgatoryData?.reason}`}
             text={purgatory.description}
             state="error"
           />
